requirements.txt
appdirs==1.4.4
asgiref==3.8.1
certifi==2025.1.31
distlib==0.3.9
Django~=5.1.7
django-mathfilters==1.0.0
filelock==3.18.0
pillow==11.1.0
pipenv==2024.4.1
platformdirs==4.3.7
psycopg2==2.9.10
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
virtualenv==20.29.3
virtualenv-clone==0.5.7
zipp==3.21.0
gunicorn==23.0.0
## The following requirements were added by pip freeze:
packaging==24.2

djangorestframework~=3.16.0
drf-yasg~=1.21.10
django-filter~=25.1
djangorestframework_simplejwt
celery~=5.5.3
requests~=2.32.4
python-dotenv~=1.1.1